feat: support configuring log level
