+++ pkg/aws/lambda.go
@@ -7,8 +7,10 @@ import (
 
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/aws/aws-lambda-go/lambdacontext"
+	"github.com/suzuki-shunsuke/slog-error/slogerr"
 	"github.com/suzuki-shunsuke/validate-pr-review-app/pkg/config"
 	"github.com/suzuki-shunsuke/validate-pr-review-app/pkg/controller"
+	"github.com/suzuki-shunsuke/validate-pr-review-app/pkg/log"
 )
 
 // Read config and secret
@@ -25,13 +27,22 @@ type Controller interface {
 	Run(ctx context.Context, logger *slog.Logger, req *controller.Request) error
 }
 
-func NewHandler(ctx context.Context, logger *slog.Logger, version string) (*Handler, error) {
+func NewHandler(ctx context.Context, logger *slog.Logger, version string, logLevel *slog.LevelVar) (*Handler, error) {
 	// read config from the environment variable
 	// parse config as YAML
 	cfg := &config.Config{}
 	if err := readConfig(cfg); err != nil {
 		return nil, err
 	}
+
+	if cfg.LogLevel != "" {
+		lvl, err := log.ParseLevel(cfg.LogLevel)
+		if err != nil {
+			return nil, fmt.Errorf("parse log level: %w", slogerr.With(err, "log_level", cfg.LogLevel))
+		}
+		logLevel.Set(lvl)
+	}
+
 	// read secrets from AWS SecretsManager
 	secret, err := readSecret(ctx, cfg.AWS.SecretID)
 	if err != nil {
