+++ pkg/config/config.go
@@ -4,7 +4,6 @@ import (
 	_ "embed"
 	"fmt"
 	"html/template"
-	"io"
 	"path"
 	"strings"
 
@@ -24,6 +23,7 @@ type Config struct {
 	UniqueUntrustedMachineUsers map[string]struct{}           `yaml:"-"`
 	Templates                   map[string]string             `yaml:"templates"`
 	BuiltTemplates              map[string]*template.Template `yaml:"-"`
+	LogLevel                    string                        `yaml:"log_level"`
 }
 
 type AWS struct {
@@ -96,61 +96,10 @@ func (c *Config) Init() error { //nolint:cyclop
 	return c.testTemplate()
 }
 
-func (c *Config) initTemplates() error {
-	defaultTemplates := map[string]string{
-		"footer":                string(templateFooter),
-		"settings":              string(templateSettings),
-		"approved":              string(templateApproved),
-		"no_approval":           string(templateNoApproval),
-		"require_two_approvals": string(templateRequireTwoApprovals),
-		"error":                 string(templateError),
-	}
-	if c.Templates == nil {
-		c.Templates = map[string]string{}
-	}
-	for name, tpl := range defaultTemplates {
-		if _, ok := c.Templates[name]; !ok {
-			c.Templates[name] = tpl
-		}
-	}
-	var define string
-	for k, v := range c.Templates {
-		define += `{{define "` + k + `"}}` + v + "{{end}}"
-	}
-	keys := []string{
-		"no_approval",
-		"approved",
-		"require_two_approvals",
-		"error",
-	}
-	templates := make(map[string]*template.Template, len(keys))
-	for _, k := range keys {
-		tpl := c.Templates[k] + define
-		tplParsed, err := template.New("_").Parse(tpl)
-		if err != nil {
-			return fmt.Errorf("parse the template %s: %w", k, err)
-		}
-		templates[k] = tplParsed
-	}
-	c.BuiltTemplates = templates
-	return nil
-}
-
 func (c *Config) testUntrustedMachineUsers() error {
 	for pattern := range c.UniqueUntrustedMachineUsers {
 		if _, err := path.Match(pattern, "foo"); err != nil {
 			return fmt.Errorf("invalid untrusted machine user pattern %q: %w", pattern, err)
-		}
-	}
-	return nil
-}
-
-func (c *Config) testTemplate() error {
-	// TODO add test cases
-	result := &Result{}
-	for key, tpl := range c.BuiltTemplates {
-		if err := tpl.Execute(io.Discard, result); err != nil {
-			return fmt.Errorf("test template %s: %w", key, err)
 		}
 	}
 	return nil
